feat(order, cart): update placeOrder

+++ src/main/java/com/example/springecom/repo/OrderRepo.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface OrderRepo extends JpaRepository<Order, Integer> {
+public interface OrderRepo extends JpaRepository<Order, Long> {
 
     List<Order> findByUser(User user);
